+++ package.json
@@ -32,8 +32,6 @@
     "babel-preset-react-app": "^9.1.2",
     "camelcase": "^5.3.1",
     "dotenv": "^8.2.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^7.2.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-plugin-import": "^2.21.2",
