+++ template.json
@@ -5,8 +5,6 @@
       "@testing-library/react": "^11.2.5",
       "@testing-library/user-event": "^12.8.3",
       "axios": "^0.21.1",
-      "enzyme-adapter-react-16": "^1.15.2",
-      "enzyme": "^3.11.0",
       "eslint-config-airbnb": "^18.2.0",
       "history": "^4.10.1",
       "node-sass": "^4.14.1",
