--- template/src/app.rtl.test.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-
-import { App } from 'app';
-
-describe('App', () => {
-  it('should render correctly', () => {
-    expect(() => render(<App />)).not.toThrow();
-  });
-});
