+++ template/src/app.test.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { render } from '@testing-library/react';
+
 import { App } from 'app';
 
 describe('App', () => {
   it('should render correctly', () => {
-    const output = shallow(<App />);
-    expect(output.exists()).toBe(true);
+    expect(() => render(<App />)).not.toThrow();
   });
 });
